Improve circle tracker behaviour
This patch improves the behaviour of the circle tracker control for the
alarm time selection. The approach is now to move the thumb to the point
on the circle that intersects the line from the centre of the circle to
where the user has touched.

To get the relevant coordinates, we assume that the svg element for the
control is the "first" element on the page. This is a bit of a hack.